Add `examples/vertex-ai/...` for LLM fine-tuning with TRL (#44)
* Add `examples/vertex-ai/trl-lora-fine-tuning`

* Add `example.py`

Easier for reproducibilty

* Update `notebook.ipynb`

* Update `example.py` and `notebook.ipynb`

* Update `images/can-it-run-llm.png`

* Add `trl-lora-sft-fine-tuning-on-vertex-ai.ipynb` example

* Move `vertex-ai/notebooks/imgs` to `vertex-ai/notebooks/assets`

* Remove notebook unrelated to Vertex AI

* Remove `gcs-empty-bucket.png` as not neeeded

* Set private `container_uri` and update `environment_variables`

Once the container is published publicly to GCR, the URL will be updated
to the one under GCR instead.

* Add `trl-full-sft-fine-tuning-on-vertex-ai.ipynb`

* Remove `TRL_USE_RICH` while it's fixed in `trl`

* Update `trl-{full,lora}-sft-fine-tuning-on-vertex-ai.ipynb`

- Remove double-quotes from environment variable definition, otherwise
it's injected and e.g. `gcloud config set project $PROJECT_ID` fails as
it will try to set the project `"my-project"` instead of `my-project`
- Replace `gsutil` command with `gcloud storage buckets create` since
`gsutil` was not working
- Remove https://github.com/alvarobartt/trl fork installation of `trl`
since it's no longer needed as those issues are already fixed within
`trl` from v0.9.6 onwards
- Set `HF_HOME` environment variable when running the SFT + LoRA
fine-tune too for consistency

* Restructure `examples/vertex-ai/notebooks`

To follow / match the same structured defined within
https://github.com/huggingface/Google-Cloud-Containers/pull/59 at
https://github.com/huggingface/Google-Cloud-Containers/pull/59/commits/30bcf7dd81b0b978555bfadcad230f7a2eec6e35

* Add missing `TRL_USE_RICH` to display progress bars

* Update `examples/vertex-ai/notebooks/trl-*` examples

Now those point to the latest publicly released PyTorch Training
container URI instead, as well as contain the
`huggingface_hub.interpreter_login` function call as the token is
required to run the training as the model is gated.