Small updates to the TGI Cloud Run readme (#105)
* Update Cloud Run TGI guide

When using Google Cloud Workstations and Cloud Shell (bash terminal), the $ACCESS_TOKEN needs to be passed in using double-quotes in the cURL command. Otherwise, you'll get error: 

```
Your client does not have permission to the requested URL <code>/v1/chat/completions</code>.
```

which might be difficult to debug since the Access Token is valid. It's just not making it through the cURL command. I do not believe using double-quotes here has any impact on using other shells.

* updated minimum CPU and Memory requirements

updated in gcloud and clarified in the description