Add `pytorch/training/gpu/2.3.1/transformers/4.48.0/py311/Dockerfile` (#134)
* Add latest PyTorch DLC with bumped dependencies

* Fix `Dockerfile` due to extra `&&`

* Lower `flash-attn` dependency version

* Add `uv` to install `pip` dependencies faster

This commit also contains some formatting improvements to better debug
the `Dockerfile` such as indentation when a command is divided in
multiple lines to know that it refers to the unindented command above;
also set bash as the default shell, and fix `gcloud` CLI installation

* Bump `transformers` to 4.48.0 and fix `Dockerfile` formatting

Bump the `transformers` dependency to 4.48.0 to support the ModernBERT
architecture, as well as bumping `diffusers` including new video and
image generation pipelines, as well as a bunch of other features,
improvements and bug fixes. Additionally, the `Dockerfile` formatting
has been fixed.

* Update `containers/pytorch/training/README.md`

* Fix `containers/pytorch/training/README.md`

* Set `transformers` version to 4.47.1 instead

* Remove `--upgrade` flag from `torch` and `transformers` install

* Bump `torch` to 2.3.1 and move `Dockerfile`

* Remove `uv` from `Dockerfile`

* Upgrade `transformers` to 4.48.0

* Remove strict version pinning on `protobuf`